+++ map/tilesets/sewer.tsx
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<tileset version="1.10" tiledversion="1.10.2" name="ENV_Tileset_Sewer" tilewidth="18" tileheight="18" tilecount="392" columns="14">
+ <image source="sewer.png" width="252" height="504"/>
+</tileset>